requirements.txt
fastapi
uvicorn
openai
chromadb
pdfplumber
whisper
ffmpeg-python
python-dotenv
tqdm
google-generativeai 